% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_data_explore.R
\name{available_icd9_versions}
\alias{available_icd9_versions}
\title{show available ICD-9 versions}
\usage{
available_icd9_versions()
}
\description{
show available ICD-9 versions
}
\examples{
\dontrun{
available_icd9_versions()
names(icd.data::icd9cm_billable)
icd9cm_vx <- icd9cm_merge_version_x(26)
}
}
