% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_data_explore.R
\name{icd9cm_get_nonbillable}
\alias{icd9cm_get_nonbillable}
\title{Function to get all onbillable ICD9 codes and labels}
\usage{
icd9cm_get_nonbillable()
}
\description{
Function to get all onbillable ICD9 codes and labels
}
