% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd9_sg_get_data_v1.R
\name{icd9cm_pcs_get_nonbillable}
\alias{icd9cm_pcs_get_nonbillable}
\title{Get all nonbillable ICD9 PROC codes and labels}
\usage{
icd9cm_pcs_get_nonbillable()
}
\description{
Get all nonbillable ICD9 PROC codes and labels
}
