% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd9_app_v2.R
\name{shine_icd9_app_v2}
\alias{shine_icd9_app_v2}
\title{Launch shiny app for v2 app}
\usage{
shine_icd9_app_v2(...)
}
\arguments{
\item{...}{Arguments passed to \code{shiny::shinyApp}}
}
\description{
Launch shiny app for v2 app
}
