% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{shine_icd_app}
\alias{shine_icd_app}
\title{Launch the shiny application}
\usage{
shine_icd_app(...)
}
\arguments{
\item{...}{Arguments passed on to \code{shiny::shinyApp}}
}
\description{
Launch the shiny application
}
