% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.icd.package.R
\docType{package}
\name{shiny.icd-package}
\alias{shiny.icd}
\alias{shiny.icd-package}
\title{Shiny interface for ICD data}
\description{
Shiny interface for ICD data
}
\author{
\strong{Maintainer}: Davit Sargsyan \email{ds752@rwjms.rutgers.edu}

Authors:
\itemize{
  \item Jack O. Wasey \email{jack@jackwasey.com}
}

}
\concept{
ICD-9 ICD-10 comorbidity comorbidities
}
